requirements.txt
pandas>=2.0.0
numpy>=1.24.0
pandasai>=2.0.0
crewai>=0.1.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
httpx>=0.24.0
langchain-openai>=0.0.5
fastapi>=0.100.0
uvicorn>=0.23.0
